requirements.txt
Django~=5.2.5
Django==5.1.5
django_environ==0.12.0
django_phonenumber_field~=8.1.0
djangorestframework~=3.16.1
djangorestframework_simplejwt==5.5.0

django-cors-headers~=4.8.0